Move volpkg conversion into vc_core; GUI calls library directly
The volpkg legacy-to-JSON conversion was previously implemented entirely
inside apps/src/vc_volpkg_convert.cpp, and the GUI launched that binary
via QProcess and parsed its stdout/stderr to surface errors. When the
binary failed (or its stderr was empty), the user saw an unresizable
QMessageBox with a truncated title and no body — no actionable info.

Move the conversion logic into vc_core (`vc::convertVolpkg` in
util/VolpkgConvert.hpp). Both local-directory and remote (s3://, https://)
inputs are handled; failures return a structured `VolpkgConvertResult`
with a human-readable `message` instead of being printed to stderr.

The CLI binary is now a thin wrapper over the library.

The GUI in MenuActionController::runLegacyVolpkgConvert now calls
`vc::convertVolpkg` directly, drops QProcess, sets a wait cursor for
the duration of the call, and shows the structured failure message
verbatim in a properly-sized message box on failure.
