+++ volume-cartographer/core/CMakeLists.txt
@@ -11,6 +11,7 @@ add_library(vc_core
         src/PointIndex.cpp
         src/Logging.cpp
         src/CrashHandler.cpp
+        src/VolpkgConvert.cpp
         src/SparseChunkSpool.cpp
         src/Slicing.cpp
         src/Compositing.cpp
